requirements.txt
Pillow==2.4.0
gevent==1.0
requests==2.2.1
